requirements.txt
Django>=4
selenium>=4
djangorestframework==3.14